-Made some changes (intending to restore the previous commit.)
