+++ src/application/LifeCalculatorView.fxml
@@ -6,19 +6,16 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.layout.VBox?>
 
-<VBox xmlns="http://javafx.com/javafx/8"
-	xmlns:fx="http://javafx.com/fxml/1"
-	fx:controller="application.LifeCalculatorController">
-	<children>
-		<HBox prefHeight="100.0" prefWidth="200.0">
-			<children>
-				<Button alignment="TOP_LEFT" mnemonicParsing="false"
-					onAction="#goToMainInputScene" text="Start Calculation">
-					<HBox.margin>
-						<Insets left="47.0" top="50.0" />
-					</HBox.margin>
-				</Button>
-			</children>
-		</HBox>
-	</children>
+<VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.LifeExpectancyCalculatorController">
+   <children>
+      <HBox prefHeight="100.0" prefWidth="200.0">
+         <children>
+            <Button mnemonicParsing="false" onAction="#goToMainInputScene" text="Start Calculation">
+               <HBox.margin>
+                  <Insets left="50.0" top="50.0" />
+               </HBox.margin>
+            </Button>
+         </children>
+      </HBox>
+   </children>
 </VBox>
