+++ src/application/LifeExpectancyCalculatorView.fxml
