+++ src/application/Main.java
@@ -23,8 +23,8 @@ public class Main extends Application {
 	public void start(Stage primaryStage) {
 		try {
 			FXMLLoader loader = new FXMLLoader();
-			VBox mainMenuContainer = loader.load(new FileInputStream("src/application/LifeExpectancyCalculatorView.fxml"));
-			LifeExpectancyCalculatorController controller = (LifeExpectancyCalculatorController)loader.getController();
+			VBox mainMenuContainer = loader.load(new FileInputStream("src/application/LifeCalculatorView.fxml"));
+			LifeCalculatorController controller = (LifeCalculatorController)loader.getController();
 			controller.setApplicationStage(primaryStage);
 			Scene mainMenuScene = new Scene(mainMenuContainer,400,400);
